requirements.txt
yfinance
pandas
pandas_ta
python-dotenv
requests
openai